@@ -1,34 +1,34 @@
 package com.github.dakusui.combinatoradix;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class EnumeratorErrorTest {
-  Enumerator.Base<String> enumerator = new Permutator<String>(Collections.singletonList("Hello"), 1);
+  Enumerator.Base<String> enumerator = new Permutator<>(Collections.singletonList("Hello"), 1);
 
-  @Test(expected = NoSuchElementException.class)
+  @Test
   public void goOverLimitByNextMethod() {
     Iterator<List<String>> i = enumerator.iterator();
     assertTrue(i.hasNext());
     assertEquals("Hello", i.next().get(0));
     assertFalse(i.hasNext());
-    i.next();
+    assertThrows(NoSuchElementException.class, i::next);
   }
 
-  @Test(expected = IndexOutOfBoundsException.class)
+  @Test
   public void goOverLimitByGetMethod() {
     assertEquals(1, enumerator.size());
-    enumerator.get(1);
+    assertThrows(IndexOutOfBoundsException.class, () -> enumerator.get(1));
   }
 
-  @Test(expected = UnsupportedOperationException.class)
+  @Test
   public void tryToRemoveAndMakeSureExceptionIsThrown() {
-    enumerator.iterator().remove();
+    assertThrows(UnsupportedOperationException.class, () -> enumerator.iterator().remove());
   }
 }
