@@ -1,6 +1,6 @@
 package com.github.dakusui.combinatoradix;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
 import java.util.Iterator;
@@ -10,7 +10,7 @@ import java.util.List;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class HomogeniousCombinatorTest extends CombinatorTestBase {
   @Test
@@ -88,7 +88,7 @@ public class HomogeniousCombinatorTest extends CombinatorTestBase {
   @Test
   public void test_empty() {
     // Empty set should result in empty iterator immediately
-    Iterator<List<String>> i = new HomogeniousCombinator<String>(Collections.<String>emptyList(), 0).iterator();
+    Iterator<List<String>> i = new HomogeniousCombinator<>(Collections.<String>emptyList(), 0).iterator();
     assertTrue(i.hasNext());
     assertEquals(emptyList(), i.next());
     assertFalse(i.hasNext());
@@ -96,65 +96,65 @@ public class HomogeniousCombinatorTest extends CombinatorTestBase {
 
   @Test
   public void test_nH1() {
-    Iterator<List<String>> homogeniousCombinator = new HomogeniousCombinator<String>(testset1(), 1).iterator();
-    assertEquals(true, homogeniousCombinator.hasNext());
+    Iterator<List<String>> homogeniousCombinator = new HomogeniousCombinator<>(testset1(), 1).iterator();
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(singletonList("A"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(singletonList("B"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(singletonList("C"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(singletonList("D"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(singletonList("E"), homogeniousCombinator.next());
-    assertEquals(false, homogeniousCombinator.hasNext());
+    assertFalse(homogeniousCombinator.hasNext());
   }
 
   @Test
   public void test_nH2() {
     Iterator<List<String>> homogeniousCombinator = Enumerators.homogeniousCombinator(testset1(), 2).iterator();
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("A", "A"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("A", "B"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("A", "C"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("A", "D"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("A", "E"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("B", "B"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("B", "C"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("B", "D"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("B", "E"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("C", "C"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("C", "D"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("C", "E"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("D", "D"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("D", "E"), homogeniousCombinator.next());
-    assertEquals(true, homogeniousCombinator.hasNext());
+    assertTrue(homogeniousCombinator.hasNext());
     assertEquals(asList("E", "E"), homogeniousCombinator.next());
-    assertEquals(false, homogeniousCombinator.hasNext());
+    assertFalse(homogeniousCombinator.hasNext());
   }
 
   public static void main(String[] args) {
-    List<String> items = new LinkedList<String>();
+    List<String> items = new LinkedList<>();
     items.add("A");
     items.add("B");
     items.add("C");
     items.add("D");
     items.add("E");
 
-    Iterator<List<String>> enumerator = new HomogeniousCombinator<String>(items, 2).iterator();
+    Iterator<List<String>> enumerator = new HomogeniousCombinator<>(items, 2).iterator();
     int i = 0;
     while (enumerator.hasNext()) {
       System.out.println(i + ":" + enumerator.next());
