@@ -2,7 +2,7 @@ package com.github.dakusui.combinatoradix;
 
 import com.github.dakusui.combinatoradix.utils.InternalUtils;
 import org.hamcrest.CoreMatchers;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
 import java.util.Iterator;
@@ -13,7 +13,7 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class PermutatorTest extends EnumeratorTestBase {
   @Test
@@ -80,7 +80,7 @@ public class PermutatorTest extends EnumeratorTestBase {
   }
 
   private void exerciseTest(List<String> symbols, int k) {
-    final Permutator<String> permutator = new Permutator<String>(
+    final Permutator<String> permutator = new Permutator<>(
         symbols,
         k
     );
@@ -116,7 +116,7 @@ public class PermutatorTest extends EnumeratorTestBase {
   @Test
   public void test_empty() {
     // Empty set should result in empty iterator immediately
-    Iterator<List<String>> i = new Permutator<String>(Collections.<String>emptyList(), 0).iterator();
+    Iterator<List<String>> i = new Permutator<>(Collections.<String>emptyList(), 0).iterator();
     assertTrue(i.hasNext());
     assertEquals(emptyList(), i.next());
     assertFalse(i.hasNext());
@@ -124,95 +124,95 @@ public class PermutatorTest extends EnumeratorTestBase {
 
   @Test
   public void test_nP0() {
-    Iterator<List<String>> p = new Permutator<String>(testset1(), 0).iterator();
-    assertEquals(true, p.hasNext());
+    Iterator<List<String>> p = new Permutator<>(testset1(), 0).iterator();
+    assertTrue(p.hasNext());
     assertEquals(emptyList(), p.next());
-    assertEquals(false, p.hasNext());
+    assertFalse(p.hasNext());
   }
 
   @Test
   public void test_nP1() {
-    Iterator<List<String>> p = new Permutator<String>(testset1(), 1).iterator();
-    assertEquals(true, p.hasNext());
+    Iterator<List<String>> p = new Permutator<>(testset1(), 1).iterator();
+    assertTrue(p.hasNext());
     assertEquals(singletonList("A"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(singletonList("B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(singletonList("C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(singletonList("D"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(singletonList("E"), p.next());
-    assertEquals(false, p.hasNext());
+    assertFalse(p.hasNext());
   }
 
   @Test
   public void test_nP2() {
-    Iterator<List<String>> p = new Permutator<String>(testset1(), 2).iterator();
-    assertEquals(true, p.hasNext());
+    Iterator<List<String>> p = new Permutator<>(testset1(), 2).iterator();
+    assertTrue(p.hasNext());
     assertEquals(asList("A", "B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("A", "C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("A", "D"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("A", "E"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("B", "A"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("B", "C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("B", "D"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("B", "E"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("C", "A"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("C", "B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("C", "D"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("C", "E"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("D", "A"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("D", "B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("D", "C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("D", "E"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("E", "A"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("E", "B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("E", "C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("E", "D"), p.next());
-    assertEquals(false, p.hasNext());
+    assertFalse(p.hasNext());
   }
 
   @Test
   public void test_nPn() {
     Iterator<List<String>> p = Enumerators.permutator(testset2(), 3).iterator();
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("A", "B", "C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("A", "C", "B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("B", "A", "C"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("B", "C", "A"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("C", "A", "B"), p.next());
-    assertEquals(true, p.hasNext());
+    assertTrue(p.hasNext());
     assertEquals(asList("C", "B", "A"), p.next());
-    assertEquals(false, p.hasNext());
+    assertFalse(p.hasNext());
   }
 
   @Test
   public void test_forEach() {
-    Enumerator.Base<String> p = new Permutator<String>(testset2(), 3);
+    Enumerator.Base<String> p = new Permutator<>(testset2(), 3);
     int i = 0;
     for (List<String> entry : p) {
       assertEquals(3, entry.size());
